Fix parsing max i64 values

+++ tests/de/builtins.rs
@@ -89,11 +89,11 @@ mod uint {
 mod int {
     mod max {
         mod unsigned {
-            test!(u8   , "int64", u8::max_value()   , u8   , u8::max_value());
-            test!(u16  , "int64", u16::max_value()  , u16  , u16::max_value());
-            test!(u32  , "int64", u32::max_value()  , u32  , u32::max_value());
-            test!(u64  , "int64", u64::max_value()  , u64  , u64::max_value());
-            test!(usize, "int"  , usize::max_value(), usize, usize::max_value());
+            test!(u8   , "int64", i8::max_value() as u8      , u8   , i8::max_value() as u8);
+            test!(u16  , "int64", i16::max_value() as u16    , u16  , i16::max_value() as u16);
+            test!(u32  , "int64", i32::max_value() as u32    , u32  , i32::max_value() as u32);
+            test!(u64  , "int64", i64::max_value() as u64    , u64  , i64::max_value() as u64);
+            test!(usize, "int"  , isize::max_value() as usize, usize, isize::max_value() as usize);
         }
 
         mod signed {
